+++ src/path.cc
@@ -133,8 +133,11 @@ std::string expand_windows(const char* p) {
   // # ~user case
   if (i != 1) {
     char* home_str = strdup(home);
+    if (home_str == nullptr) {
+      Rf_error("Allocation Failed");
+    }
     strncpy(home, dirname(home_str), PATH_MAX);
-    delete[] home_str;
+    free(home_str);
 
     // only copy enough characters to i
     n = strlen(home);