+++ NEWS.md
@@ -5,6 +5,9 @@
 * `fs_path` `fs_bytes` and `fs_perm` objects now use `methods::setOldClass()`
   so that S4 dispatch to their base classes works as intended (#91).
 
+* Fix allocation bug in `path_exists()` using `delete []` when we should have
+  used `free()`.
+
 # fs 1.2.1
 
 ## Features